refactor standup recording for more consistency
