+++ lib/helpers/index.js
@@ -3,5 +3,8 @@
 module.exports = {
   time: require('./time'),
   getStandupReport: require('./getStandupReport'),
-  getUser: require('./getUser')
+  getUser: require('./getUser'),
+  doBlock: require('./doBlock'),
+  doChannelReport: require('./doChannelReport'),
+  doInterview: require('./doInterview')
 };
